激励网站 11/10 提交

+++ src/main/java/com/astrosea/richer/vo/BaseRewVo.java
@@ -2,9 +2,11 @@ package com.astrosea.richer.vo;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class BaseRewVo {
 
-    private Long rewBase;
+    private BigDecimal rewBase;
 
 }
